ccai_insights_function/requirements.txt
google-cloud-contact-center-insights